requirements.txt
Django==3.2.3
psycopg2-binary>=2.8
GitPython==3.1.17